pyproject.toml
[project]
name = "issue-grouping"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "apscheduler==3.11.0",
    "black>=25.1.0",
    "faiss-cpu>=1.12.0",
    "fastapi>=0.116.1",
    "ipykernel>=6.30.1",
    "isort>=6.0.1",
    "langchain>=0.3.27",
    "langchain-community>=0.3.29",
    "mysql-connector-python==9.4.0",
    "nest-asyncio>=1.6.0",
    "openpyxl==3.1.5",
    "qgenie-sdk[all]>=0.2.2",
    "rapidfuzz===3.14.0",
    "scikit-learn>=1.7.1",
    "streamlit==1.47.0",
    "uvicorn[standard]>=0.35.0",
]

[[tool.uv.index]]
url = "https://devpi.qualcomm.com/qcom/dev/+simple"
default = true
